+++ daemon/main.rs
@@ -27,14 +27,6 @@ mod task_handler;
 
 #[async_std::main]
 async fn main() -> Result<()> {
-    // Get settings from the configuration file and the program defaults.
-    let settings = Settings::new()?;
-    // Immediately save it. This also creates the save file in case it didn't exist yet.
-    if let Err(error) = settings.save() {
-        println!("Failed saving config file.");
-        println!("{:?}", error);
-    }
-
     // Parse commandline options.
     let opt = Opt::from_args();
 
@@ -42,7 +34,7 @@ async fn main() -> Result<()> {
         fork_daemon(&opt)?;
     }
 
-    // Set the verbosity level for the client app.
+    // Set the verbosity level of the logger.
     if opt.verbose >= 3 {
         SimpleLogger::init(LevelFilter::Debug, Config::default())?;
     } else if opt.verbose == 2 {
@@ -52,6 +44,25 @@ async fn main() -> Result<()> {
     } else if opt.verbose == 0 {
         SimpleLogger::init(LevelFilter::Error, Config::default())?;
     }
+
+    // Try to read settings from the configuration file.
+    let settings = match Settings::read(false) {
+        Ok(settings) => settings,
+        Err(_) => {
+            // There's something wrong with the config file or something's missing.
+            // Try to read the config and fill missing values with defaults.
+            // This might be possible on version upgrade or first run.
+            let settings = Settings::new(false)?;
+
+            // Since we needed to add values to the configuration, we have to save it.
+            // This also creates the save file in case it didn't exist yet.
+            if let Err(error) = settings.save() {
+                println!("Failed saving config file.");
+                println!("{:?}", error);
+            }
+            settings
+        }
+    };
 
     init_directories(&settings.shared.pueue_directory);