+++ shared/settings.rs
@@ -2,8 +2,9 @@ use std::collections::HashMap;
 use std::fs::{create_dir_all, File};
 use std::io::prelude::*;
 
-use anyhow::{anyhow, Result};
+use anyhow::{anyhow, bail, Result};
 use config::Config;
+use log::info;
 use rand::Rng;
 use serde_derive::{Deserialize, Serialize};
 
@@ -30,14 +31,9 @@ pub struct Client {
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Daemon {
     pub default_parallel_tasks: usize,
-    #[serde(default = "pause_on_failure_default")]
     pub pause_on_failure: bool,
     pub callback: Option<String>,
     pub groups: HashMap<String, usize>,
-}
-
-fn pause_on_failure_default() -> bool {
-    false
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -53,7 +49,9 @@ impl Settings {
     /// If a local config file already exists it is parsed and
     /// overwrites the default option values.
     /// The local config is located at "~/.config/pueue.yml".
-    pub fn new() -> Result<Settings> {
+    ///
+    /// If `require_config` is `true`, an error will be thrown, if no configuration file can be found.
+    pub fn new(require_config: bool) -> Result<Settings> {
         let mut config = Config::new();
 
         config.set_default("shared.port", "6924")?;
@@ -73,7 +71,21 @@ impl Settings {
         config.set_default("daemon.groups", HashMap::<String, i64>::new())?;
 
         // Add in the home config file
-        parse_config(&mut config)?;
+        parse_config(&mut config, require_config)?;
+
+        // Try to can deserialize the entire configuration
+        Ok(config.try_into()?)
+    }
+
+    /// Try to read the config file without any default values.
+    /// This is done by the daemon on startup.
+    /// If the file can be read without any need for defaults, we don't have to persist it
+    /// afterwards.
+    pub fn read(require_config: bool) -> Result<Settings> {
+        let mut config = Config::new();
+
+        // Merge configuration files we can find in ascending order.
+        parse_config(&mut config, require_config)?;
 
         // Try to can deserialize the entire configuration
         Ok(config.try_into()?)
@@ -103,16 +115,24 @@ impl Settings {
 /// Get all possible configuration paths and check if there are
 /// configuration files at those locations.
 /// All configs will be merged by importance.
-fn parse_config(settings: &mut Config) -> Result<()> {
-    //println!("Parsing config files");
+///
+/// If `require_config` is `true`, an error will be thrown, if no configuration file can be found.
+fn parse_config(settings: &mut Config, require_config: bool) -> Result<()> {
+    let mut config_found = false;
+    info!("Parsing config files");
     for directory in get_config_directories()?.into_iter() {
         let path = directory.join("pueue.yml");
-        //println!("Checking path: {:?}", &path);
+        info!("Checking path: {:?}", &path);
         if path.exists() {
-            //println!("Parsing config file at: {:?}", path);
+            info!("Found config file at: {:?}", path);
+            config_found = true;
             let config_file = config::File::with_name(path.to_str().unwrap());
             settings.merge(config_file)?;
         }
+    }
+
+    if require_config && !config_found {
+        bail!("Couldn't find a configuration file. Did you start the daemon yet?");
     }
 
     Ok(())